+++ src/install.ts
@@ -50,7 +50,7 @@ async function getOsVersion(): Promise<string | undefined> {
     return match?.[1];
   }
 
-  if (runnerOs === 'MacOS') {
+  if (runnerOs === 'macOS') {
     const output = await exec.getExecOutput('sw_vers', ['-productVersion'], {
       silent: true,
     });
@@ -77,15 +77,20 @@ async function getCacheKey(
   version: ResolvedVersion,
 ): Promise<string> {
   const runnerOs = process.env.RUNNER_OS;
+  const runnerArch = process.env.RUNNER_ARCH;
   const jobId = process.env.GITHUB_JOB;
   const osVersion = await getOsVersion();
 
-  if (runnerOs === undefined || jobId === undefined) {
-    core.setFailed('Could not determine runner OS or job ID');
+  if (
+    runnerOs === undefined ||
+    runnerArch === undefined ||
+    jobId === undefined
+  ) {
+    core.setFailed('Could not determine runner OS, runner arch or job ID');
     process.exit(1);
   }
 
-  let hashKey = jobId + runnerOs + (osVersion ?? '');
+  let hashKey = jobId + runnerOs + runnerArch + (osVersion ?? '');
 
   hashKey += input.source.type;
   if (input.source.type === 'registry') {